chatbot/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-genai==1.0.0
python-dotenv==1.0.0
fastmcp==0.1.0
aiohttp==3.8.5
